backend/requirements.txt
# Core Framework
fastapi>=0.115.6
uvicorn[standard]>=0.34.0
python-dotenv>=1.0.1
pydantic>=2.10.4
pydantic-settings>=2.7.1
sqlalchemy>=2.0.36
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.1

# LangChain + LangGraph
langchain>=0.3.14
langchain-openai>=0.3.0
langchain-community>=0.3.14
langgraph>=0.2.60

# RAG Pipeline
chromadb>=0.5.23
pymupdf>=1.25.3

# Utilities
python-multipart>=0.0.20
websockets>=14.1
httpx>=0.28.1
aiofiles>=24.1.0

# Testing
pytest>=8.3.4
pytest-asyncio>=0.25.0
